adding changes

+++ src/test/java/com/github/curriculeon/RandomUtilsTest.java
@@ -106,7 +106,7 @@ public class RandomUtilsTest {
         for (int i = 0; i < numberOfIterations; i++) {
             // create a random string
             String randomString = RandomUtils.createString(min, max, stringLength);
-            Assert.assertEquals(stringLength, randomString.length());
+            Assert.assertEquals(stringLength,(Integer) randomString.length());
 
             // check each character of string
             for (char character : randomString.toCharArray()) {
@@ -127,11 +127,11 @@ public class RandomUtilsTest {
         for (int i = 0; i < numberOfIterations; i++) {
             // create a random string array
             String[] randomStrings = RandomUtils.createStrings(min, max, stringLength, numberOfStrings);
-            Assert.assertEquals(numberOfStrings, randomStrings.length);
+            Assert.assertEquals(numberOfStrings, (Integer) randomStrings.length);
             for (String randomString : randomStrings) {
                 // check each character of string
 
-                Assert.assertEquals(stringLength, randomString.length());
+                Assert.assertEquals(stringLength, (Integer) randomString.length());
                 for (char character : randomString.toCharArray()) {
                     boolean outcome = RangeChecker.isInRange(character, min, max);
                     Assert.assertTrue(outcome);
